refactor(styles): use public withWrapper
