+++ src/link/index.js
@@ -8,7 +8,7 @@ LICENSE file in the root directory of this source tree.
 // Styled elements
 import * as React from 'react';
 import {Link} from './styled-components.js';
-import {withWrapper} from '../styles/styled.js';
+import {withWrapper} from '../styles/index.js';
 import {isFocusVisible, forkFocus, forkBlur} from '../utils/focusVisible.js';
 
 function LinkFocus(props) {
