+++ src/phone-input/styled-components.js
@@ -7,7 +7,7 @@ LICENSE file in the root directory of this source tree.
 // @flow
 import * as React from 'react';
 import {SIZE} from './constants.js';
-import {styled, withStyle} from '../styles/index.js';
+import {styled, withStyle, withWrapper} from '../styles/index.js';
 import {StyledList} from '../menu/index.js';
 import {
   StyledDropdownListItem,
@@ -15,7 +15,6 @@ import {
 } from '../select/index.js';
 import defaultProps from '../select/default-props.js';
 import type {SizeT} from './types.js';
-import {withWrapper} from '../styles/styled.js';
 
 type SizeStyleProps = {
   $size?: SizeT,
